Assert added
